+++ packages/web/src/services/FileDownloadServiceWeb.ts
@@ -1,14 +1,17 @@
+import axios from "axios";
 import StreamSaver from "streamsaver";
+
 import {
     FileDownloadService,
     DownloadResult,
     FileInfo,
     DownloadResolution,
-    HttpServiceBase,
+    FileStorageServiceBase,
 } from "../../../core/services";
-import axios from "axios";
 
-export default class FileDownloadServiceWeb extends HttpServiceBase implements FileDownloadService {
+export default class FileDownloadServiceWeb
+    extends FileStorageServiceBase
+    implements FileDownloadService {
     isFileSystemAccessible = false;
 
     public async download(fileInfo: FileInfo): Promise<DownloadResult> {
@@ -29,15 +32,6 @@ export default class FileDownloadServiceWeb extends HttpServiceBase implements F
 
         const localDownloadResult = this.isLocalPath(fileInfo.path, fileInfo);
         return localDownloadResult; // This will be either a DownloadResult or null
-    }
-
-    private isS3Url(url: string): boolean {
-        try {
-            const { protocol, hostname } = new URL(url);
-            return protocol === "https:" && hostname.endsWith(".amazonaws.com");
-        } catch (error) {
-            return false;
-        }
     }
 
     private isLocalPath(filePath: string, fileInfo: FileInfo): DownloadResult | null {
@@ -135,32 +129,6 @@ Please navigate to this directory manually, or upload files to a remote address 
         } finally {
             URL.revokeObjectURL(downloadUrl);
         }
-    }
-
-    private parseS3Url(url: string): { bucket: string; key: string; region: string } {
-        const { hostname, pathname } = new URL(url);
-        const [bucket] = hostname.split(".");
-        const key = pathname.slice(1);
-        const region = hostname.split(".")[2];
-        return { bucket, key, region };
-    }
-
-    private async listS3Objects(bucket: string, prefix: string, region: string): Promise<string[]> {
-        const url = `https://${bucket}.s3.${region}.amazonaws.com?list-type=2&prefix=${encodeURIComponent(
-            prefix
-        )}`;
-        const response = await axios.get(url);
-        const parser = new DOMParser();
-        const xmlDoc = parser.parseFromString(response.data, "text/xml");
-
-        const keys: string[] = [];
-        const contents = xmlDoc.getElementsByTagName("Key");
-
-        for (let i = 0; i < contents.length; i++) {
-            keys.push(contents[i].textContent || "");
-        }
-
-        return keys;
     }
 
     public async prepareHttpResourceForDownload(url: string, postBody: string): Promise<Blob> {
